refactor(winner-modal-ui-btn): add ui button for winner modal
